requirements.txt
django
groq 
python-dotenv
mongoengine 
dnspython
werkzeug
pymongo 
djongo
pytz
google-auth
google-auth-httplib2
requests